Common width-related code.
