+++ src/types/style.ts
@@ -128,3 +128,88 @@ export type RadiusProp = {
 };
 
 export type BackgroundColor = 'primary' | 'primaryDarken' | 'accent' | 'accentDarken' | 'alert' | 'gray' | 'white';
+
+export type CSSVariable = `var(--${string})`;
+
+// ref: https://developer.mozilla.org/en-US/docs/Web/CSS/flex-basis
+export type CSSLength =
+  | `${string}cap`
+  | `${string}ch`
+  | `${string}em`
+  | `${string}ex`
+  | `${string}ic`
+  | `${string}lh`
+  | `${string}rcap`
+  | `${string}rem`
+  | `${string}rex`
+  | `${string}ric`
+  | `${string}rlh`
+  | `${string}vh`
+  | `${string}vmax`
+  | `${string}vmin`
+  | `${string}vw`
+  | `${string}vb`
+  | `${string}vi`
+  | `${string}cqw`
+  | `${string}cqh`
+  | `${string}cqi`
+  | `${string}cqb`
+  | `${string}cqmin`
+  | `${string}cqmax`
+  | `${string}px`
+  | `${string}cm`
+  | `${string}mm`
+  | `${string}q`
+  | `${string}in`
+  | `${string}pc`
+  | `${string}pt`;
+
+export type CSSPercentage = `${string}%`;
+
+export type CSSLengthPercentage = CSSLength | CSSPercentage;
+
+export type CSSWitdh =
+  | CSSLength
+  | CSSPercentage
+  | 'auto'
+  | 'fit-content'
+  | `fit-content(${CSSLengthPercentage})`
+  | 'min-content'
+  | 'max-content'
+  | CSSVariable;
+
+export type CSSMaxWidth =
+  | 'none'
+  | CSSLengthPercentage
+  | 'min-content'
+  | 'max-content'
+  | 'fit-content'
+  | `fit-content(${CSSLengthPercentage})`
+  | CSSVariable;
+
+export type CSSMinWidth =
+  | 'auto'
+  | CSSLengthPercentage
+  | 'min-content'
+  | 'max-content'
+  | 'fit-content'
+  | `fit-content(${CSSLengthPercentage})`
+  | CSSVariable;
+
+export type WidthProps = {
+  /**
+   * 幅を指定
+   * @defaultValue auto
+   */
+  width?: CSSWitdh;
+  /**
+   * 最小幅
+   * @defaultValue auto
+   */
+  minWidth?: CSSMinWidth;
+  /**
+   * 最大幅
+   * @defaultValue none
+   */
+  maxWidth?: CSSMaxWidth;
+};
